+++ docs/release_notes.rst
@@ -8,6 +8,7 @@ Release Notes
 * Added ``JavetVirtualObject``
 * Updated ``JavetUniversalProxyHandler`` to allow passing ``V8Value``
 * Updated ``JavetUniversalProxyHandler`` to allow passing ``V8ValueFunction`` as anonymous function
+* Updated ``JavetUniversalProxyHandler`` to allow passing ``V8ValueObject`` as anonymous object
 * Added error code 602 and 603
 
 0.9.9